requirements.txt
streamlit>=1.29
opencv-python-headless
numpy
Pillow
